Added error handling.

+++ src/tasks/delete-expired-files.mjs
@@ -25,14 +25,23 @@ export function deleteExpiredFiles(pathSource, expirationInDays) {
     );
   });
 
-  filterFilesList.forEach((fileName) => {
-    const fullName = file.getFullPath(pathSource, fileName);
-    if (file.fileExists(fullName)) {
-      // TODO check error
-      file.deleteFile(fullName);
-      logFile.writeMessage(`  ${fileName} deleted.`, logFile.TYPE_MESSAGE_INFO);
-    }
-  });
+  if (filterFilesList.length > 0) {
+    filterFilesList.forEach((fileName) => {
+      const fullName = file.getFullPath(pathSource, fileName);
+      if (file.fileExists(fullName)) {
+        try {
+          file.deleteFile(fullName);
+          logFile.writeMessage(
+            `  ${fileName} deleted.`,
+            logFile.TYPE_MESSAGE_INFO
+          );
+        } catch (err) {
+          if (err.type === "delete")
+            logFile.writeMessage(`Unable to delete file ${err.file}.`);
+        }
+      }
+    });
+  } else logFile.writeMessage("No expired files.", logFile.TYPE_MESSAGE_INFO);
 
   logFile.writeMessage("Garbage finish.", logFile.TYPE_MESSAGE_SYST);
 }
